Add  breadcrumb to session review
