+++ src/main/kotlin/pl/edu/uj/ii/ksi/mordor/model/FileType.kt
@@ -1,0 +1,8 @@
+package pl.edu.uj.ii.ksi.mordor.model
+
+enum class FileType(val previewType: String) {
+    IMAGE("IMAGE"),
+    PAGE("PAGE"),
+    TOO_LARGE("TOO_LARGE"),
+    CODE("CODE")
+}
