+++ src/main/kotlin/pl/edu/uj/ii/ksi/mordor/model/RelativeDirectory.kt
@@ -1,0 +1,6 @@
+package pl.edu.uj.ii.ksi.mordor.model
+
+data class RelativeDirectory(
+    val name: String,
+    var path: String
+)
