+++ src/main/resources/templates/components/breadcrumb.html
@@ -1,0 +1,8 @@
+<div th:fragment="breadcrumb (path, dir)" th:remove="tag">
+    <ol class="breadcrumb">
+        <li th:each="dir, itrStat : ${path}" th:class="${itrStat.last}? 'breadcrumb-item active' : 'breadcrumb-item'">
+            <a th:unless="${itrStat.last}" th:href="${dir.path}" th:text="${dir.name}">directory</a>
+            <span th:if="${itrStat.last}" th:text="${dir.name}">current directory</span>
+        </li>
+    </ol>
+</div>
