+++ src/main/resources/templates/review/preview.html
@@ -18,12 +18,7 @@
         </div>
     </div>
 
-    <ol class="breadcrumb">
-        <li th:each="dir, itrStat : ${path}" th:class="${itrStat.last}? 'breadcrumb-item active' : 'breadcrumb-item'">
-            <a th:unless="${itrStat.last}" th:href="${dir.path}" th:text="${dir.name}">directory</a>
-            <span th:if="${itrStat.last}" th:text="${dir.name}">current directory</span>
-        </li>
-    </ol>
+    <div th:replace="components/breadcrumb :: breadcrumb (path=${path}, dir=${dir})"/>
 
     <th:block th:switch="${type}">
         <div th:case="'IMAGE'">
