+++ src/main/resources/templates/tree.html
@@ -6,12 +6,8 @@
 </head>
 <body>
 <div layout:fragment="content">
-    <ol class="breadcrumb">
-        <li th:each="dir, itrStat : ${path}" th:class="${itrStat.last}? 'breadcrumb-item active' : 'breadcrumb-item'">
-            <a th:unless="${itrStat.last}" th:href="${dir.path}" th:text="${dir.name}">directory</a>
-            <span th:if="${itrStat.last}" th:text="${dir.name}">current directory</span>
-        </li>
-    </ol>
+
+    <div th:replace="components/breadcrumb :: breadcrumb (path=${path}, dir=${dir})"/>
 
     <div th:if="${children.empty}">
         <div>There are no files here.</div>
